requirements.txt
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
numpy==1.21.4
olefile==0.46
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing_extensions==4.0.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
